Update MariaDB image to latest LTS
Previously, the image was pinned to 10.5 because later versions are no longer fully compatible with older MySQL drivers.

It's [End of Life Date](<https://endoflife.date/mariadb>) is approaching in a few months though; if you need full MySQL compatibility for any reason, e.g. if you re-use this database as a Database Host for a node and have a service that needs it, you can always use a MySQL image.